Add healthtest functionality to automatically exclude down hosts
Add a healthtest function activated by setting the 'test' attribute on a label.
When a test is specified, each RR within the label (A and AAAA only at this
stage) is polled regularly with a configurable test. Current configurable
tests are that a tcp port can be opened, or that an NTP response within a
given stratum range is achieved. RRs that fail the test will be excluded
from any results. Only one health test is performed per A or AAAA record
in the system (for any given healthtest configuration), i.e. multiple
RRs in one zone or multiple labels across zones will not cause testing
of the same A or AAAA record provided the tests configured by each RR
or each zone have the same parameter.

There are two categories of health test. Local health tests are run
independently of each other and simply test the health of one A or AAAA
record. Global health tests have (for any given set of testing parameters)
a single view of the status of all A and AAAA records tested by that
health test, and extract the health status of any given record from that
health test (defaulting as appropriate if it is not provided).

The following types of health test are supported:

    Type     Category   Description
    ====     ========   ===========
    tcp      local      test whether a TCP port is open
    ntp      local      test whether NTP is running at a given stratum
    exec     local      test whether a command returns a zero status
    file     global     test entries in a JSON file
    nodeping global     use Nodeping's API to get health status
    pingdom  global     use Pingdom's API to get health status

The configuration of each health test is controlled by parameters that
are common to each health test, and parameters that are specific to
each health test.

Common parameters
=================

* type: specifies type of test (currently "ntp" or "tcp")

* frequency: specifies time in seconds between polls if the server is up

* retry_time: specifies time in seconds between polls if a poll fails

* retries: number of failed polls required to consider a server as down

* timeout: timeout on each of the polls

* healthy_initially: set to true to default the status of the test to healthy

These may each be ommited and have sensible defaults

Specific parameters for type 'ntp'
==================================

* max_stratum: maximum ntp stratum number for the poll
  to be considered successful

  example configuration:

     "test" :  {
         "type" : "ntp",
         "frequency" : 30,
         "retry_time" : 5,
         "retries" : 3,
         "timeout" : 5,
         "max_stratum" : "3"
      }

Specific parameters for type 'tcp'
==================================

* port: tcp port number to connect to

    example configuration

     "test" :  {
         "type" : "tcp",
         "frequency" : 30,
         "retry_time" : 5,
         "retries" : 3,
         "timeout" : 5,
         "port" : 80
      }

Specific parameters for type 'exec'
===================================

* cmd: command to run. '{}' substituted by IP address

    example configuration

     "test" :  {
         "type" : "exec",
         "frequency" : 30,
         "retry_time" : 5,
         "retries" : 3,
         "timeout" : 5,
         "cmd" : "/bin/ping -c1 -W5 {}"
      }

Specific parameters for type 'file'
===================================

* path: JSON file to load

  the file has the format:
     { "192.200.0.1": true, "192.200.0.2": false }

    example configuration

     "test" :  {
         "type" : "file",
         "frequency" : 30,
         "retry_time" : 5,
         "retries" : 3,
         "timeout" : 5,
         "path" : "/path/to/my/file.json"
      }

Hosts not mentioned in the JSON file default to the status of
healthy_initially

Specific parameters for type 'nodeping'
=======================================

* token: Nodeping API key to use

    example configuration

     "test" :  {
         "type" : "nodeping",
         "frequency" : 30,
         "retry_time" : 5,
         "retries" : 3,
         "timeout" : 5,
         "token" : "abcdef01234567890"
      }

If the token is not specifed, it will default to a value from the
configuration file. Thus if the configuration file contains:

[nodeping]
token=abcdef01234567890

then the following much simpler configuration might be all that is
required:

    "test": {"type": "nodeping"}

Note that the label of each test in Nodeping should correspond
to the IP address being tested.

As Nodeping does not support listing of status of 'up' hosts,
hosts not mentioned in the JSON file default to the status of
'true' (i.e. healthy) all the time.

Specific parameters for type 'pingdom'
=======================================

* username: username to use with the Pingdom service

* password: password to use with the Pingdom service

* account_email: account email to use with Pingdom service (multi user accounts only)

* app_key: application key to use with the Pingdom service (has a sensible default)

* state_map: JSON map of Pingdom status to health values (e.g. true/false)

    example configuration

     "test" :  {
         "type" : "pingdom",
         "frequency" : 30,
         "retry_time" : 5,
         "retries" : 3,
         "timeout" : 5,
         "username" : "foo@example.com",
         "password" : "sekr1t",
         "app_key" : "gyxtnd2fzco8ys29m8luk4syag4ybmc0"
      }

If any of the parameters is not specifed, it will default to a value from the
configuration file. Thus if the configuration file contains:

[pingdom]
username=foo@example.com
password=serk1t
app_key=gyxtnd2fzco8ys29m8luk4syag4ybmc0

then the following much simpler configuration might be all that is
required:

    "test": {"type": "pingdom"}

Note that the name of each test in Pingdom should correspond
to the IP address being tested.

Pingdom supports more states than just 'up' and 'down'. By default
'up' is mapped to healthy (true), and 'down' or 'paused' to unhealthy
(false), with the remaining states being mapped to the default value
specified by initially_healthy. This may be customised using the
JSON state_map in either the configuration file or the test definition.

Signed-off-by: Alex Bligh <alex@alex.org.uk>

Conflicts:
	picker.go