Only perform one health test per IP / parameter tuple
Prior to this commit, if multiple A records (in the same zone or different
zones) carried the same IP address and the same test parameters, they would
still be tested multiple times. This caused unnecessary tests to run when
(e.g.) a number of virtual web site A records pointed at the same IP address,
or a number of load balancer A records pointed at the same set of load
balancers.

This commit ensures that for any given set of test parameters, only one
test is run per IP address.

It also fixes a race condition destroying test handlers which could otherwise
cause a panic due to a send on a closed channel.

Signed-off-by: Alex Bligh <alex@alex.org.uk>