Skip unchanged zonefiles
Do not reload zonefiles which are entirely unchanged as determined
by SHA256 hash. In an environment where (e.g.) zones are PUT using
a REST API the same zone may be delivered in a situation where it
is entirely unchanged. In this circumstance is it useful to avoid
reloading it.

Signed-off-by: Alex Bligh <alex@alex.org.uk>