Remove minimum TTL for NS records
geodns adopted a seemingly arbitrary minimum TTL for NS records
of 86400 seconds (one day). I can find no reason for this in the
RFCs. Setting a lower TTL is useful e.g. when NS records are
to be changed, for instance removal of a nameserver. This patch
removes this arbitrary limit.

This also removes a peculiarity where TTL on NS records would
not default to the label TTL but would default to the zone TTL.
Now TTL on NS records defaults to the label TTL but not the
zone TTL. This is because it's not possible to the set the
TTL on an RR basis.

Signed-off-by: Alex Bligh <alex@alex.org.uk>